♻️ Refactor: Main에서 Home으로 변경
